requirements.txt
google-generativeai
speechrecognition
pyaudio
requests
pydub